expose calculateCountdownState util
