+++ src/lib/drop/index.ts
@@ -1,2 +1,2 @@
+export * from './calculate-countdown-state'
 export * from './provider'
-export * from './utils'
